Host logos locally
Closes #189.
Closes #207.

This was quite a bit of a faff. However, our total file size for logos
is now down from 1400K to 604K, and image serving is much more reliable.
:)

The new process introduces a bit more manual work into the onboarding
process on our side (e.g. logo preparation), but I think this was
unavoidable all along, and we're automating things as much as is
realistic.

Changes included:

* Update member onboarding procedure in `CONTRIBUTING.md`. It has
  changed quite a bit.
* Add more instructions to `CONTRIBUTING.md` regarding logo asset
  management, license files, and others.
* Update `members.ts` accordingly.
* Deprecate `urlSquareLogoWithBackground` in
  `contrib/example-schema.json` and `src/content/config.ts`.
* Remove `urlSquareLogoWithBackground` from Join page.
* Update branding submission instructions on Join page.
* Add manually compressed and adjusted member logos to
  `public/images/members/<slug>/logo.webp`.
* Reorganise homepage logos in
  `src/assets/images/members/<slug>/minimal.svg`.
* Add license information for all images to bring us back to REUSE
  compliance.
* Remove `pixee.json.lock`, because we're not using the bad logo from
  that file anymore.